pyproject.toml
[project]
name = "querygpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.0.10",
    "langchain-google-vertexai>=2.0.21",
    "langchain-neo4j>=0.4.0",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.8",
    "mem0ai>=0.1.94",
    "qdrant-client>=1.14.2",
    "rank-bm25>=0.2.2",
]
